Add primitive types

+++ src/types/animation.rs
@@ -1,4 +1,5 @@
 use crate::types::photo_size::PhotoSize;
+use crate::types::primitive::Integer;
 
 /// This object represents an animation file (GIF or H.264/MPEG-4 AVC video without sound).
 #[derive(Debug)]
@@ -6,11 +7,11 @@ pub struct Animation {
     /// Unique file identifier
     pub file_id: String,
     /// Animation width as defined by sender
-    pub width: i64,
+    pub width: Integer,
     /// Animation height as defined by sender
-    pub height: i64,
+    pub height: Integer,
     /// Duration of the video in seconds as defined by sender
-    pub duration: i64,
+    pub duration: Integer,
     /// Animation thumbnail as defined by sender
     pub thumb: Option<PhotoSize>,
     /// Original animation filename as defined by sender
@@ -18,5 +19,5 @@ pub struct Animation {
     /// MIME type of the file as defined by sender
     pub mime_type: Option<String>,
     /// File size
-    pub file_size: Option<i64>,
+    pub file_size: Option<Integer>,
 }
