+++ src/types/file.rs
@@ -1,3 +1,5 @@
+use crate::types::primitive::Integer;
+
 /// This object represents a file ready to be downloaded.
 /// The file can be downloaded via the link https://api.telegram.org/file/bot<token>/<file_path>.
 /// It is guaranteed that the link will be valid for at least 1 hour.
@@ -8,7 +10,7 @@ pub struct File {
     /// Unique identifier for this file
     pub file_id: String,
     /// File size, if known
-    pub file_size: Option<i64>,
+    pub file_size: Option<Integer>,
     /// File path.
     /// Use https://api.telegram.org/file/bot<token>/<file_path> to get the file.
     pub file_path: Option<String>,
