+++ src/types/game_high_score.rs
@@ -1,12 +1,13 @@
+use crate::types::primitive::Integer;
 use crate::types::user::User;
 
 /// This object represents one row of the high scores table for a game.
 #[derive(Debug)]
 pub struct GameHighScore {
     /// Position in high score table for the game
-    pub position: i64,
+    pub position: Integer,
     /// User
     pub user: User,
     /// Score
-    pub score: i64,
+    pub score: Integer,
 }
