+++ src/types/location.rs
@@ -1,8 +1,10 @@
+use crate::types::primitive::Float;
+
 /// This object represents a point on the map.
 #[derive(Debug)]
 pub struct Location {
     /// Longitude as defined by sender
-    pub longitude: f64,
+    pub longitude: Float,
     /// Latitude as defined by sender
-    pub latitude: f64,
+    pub latitude: Float,
 }
