+++ src/types/mask_position.rs
@@ -1,3 +1,5 @@
+use crate::types::primitive::Float;
+
 /// This object describes the position on faces where a mask should be placed by default.
 #[derive(Debug)]
 pub struct MaskPosition {
@@ -11,14 +13,14 @@ pub struct MaskPosition {
     /// For example, choosing -1.0
     /// will place mask just
     /// to the left of the default mask position.
-    pub x_shift: f64,
+    pub x_shift: Float,
     /// Shift by Y-axis measured
     /// in heights of the mask scaled to the face size,
     /// from top to bottom.
     /// For example, 1.0 will place
     /// the mask just below the default mask position.
-    pub y_shift: f64,
+    pub y_shift: Float,
     /// Mask scaling coefficient.
     /// For example, 2.0 means double size.
-    pub scale: f64,
+    pub scale: Float,
 }
