+++ src/types/photo_size.rs
@@ -1,12 +1,14 @@
+use crate::types::primitive::Integer;
+
 /// This object represents one size of a photo or a file / sticker thumbnail.
 #[derive(Debug)]
 pub struct PhotoSize {
     /// Unique identifier for this file
     pub file_id: String,
     /// Photo width
-    pub width: i64,
+    pub width: Integer,
     /// Photo height
-    pub height: i64,
+    pub height: Integer,
     /// File size
-    pub file_size: Option<i64>,
+    pub file_size: Option<Integer>,
 }
