+++ src/types/sticker.rs
@@ -1,5 +1,6 @@
 use crate::types::mask_position::MaskPosition;
 use crate::types::photo_size::PhotoSize;
+use crate::types::primitive::Integer;
 
 /// This object represents a sticker.
 #[derive(Debug)]
@@ -7,9 +8,9 @@ pub struct Sticker {
     /// Unique identifier for this file
     pub file_id: String,
     /// Sticker width
-    pub width: i64,
+    pub width: Integer,
     /// Sticker height
-    pub height: i64,
+    pub height: Integer,
     /// Sticker thumbnail in the .webp or .jpg format
     pub thumb: Option<PhotoSize>,
     /// Emoji associated with the sticker
@@ -19,5 +20,5 @@ pub struct Sticker {
     /// For mask stickers, the position where the mask should be placed
     pub mask_position: Option<MaskPosition>,
     /// File size
-    pub file_size: Option<i64>,
+    pub file_size: Option<Integer>,
 }
